More complete initialization of AuthorizationSet
elems_capacity_ was not always initialized, though it would later be
inside Reinitialize.
This was triggering UBSan due to the size overflow in the
memset_s(nullptr, ... inside Clear()

I more completely initialized the structure while I was adding
initialization of elems_capacity_, just to live in a more predictable
world.

Bug: 129300035
Test: KM VTS 4 under Trusty with UBSan and ImportWrappedKey CL
Change-Id: Iab5b6d6e94a59a8cfc1dfe69e44f7c076baf551e