Don't construct the RP context for KM <= 4.1
The remote provisioning context constructor being called for KM devices
that don't even support RPK, requiring additional libraries to be
needlessly linked just to support this. This gates the constructor call
on whether or not the KM version is actually KeyMint 1.0 or greater.

Ignore-AOSP-First: CP from AOSP-master to sc-dev
Bug:191740898
Test: Building system/keymaster works without additional libraries for
KM 4.1

Change-Id: I0a64daf3d0aa122445612f29286e3be010315a16