Don't fail on TAG_ALLOW_WHILE_ON_BODY
The TAG_ALLOW_WHILE_ON_BODY authorization is not required to be
supported, and if it is not supported, it's a noop. Don't report a
failure in the software keyblob implementation if this tag is received
as an input authorization.

Test: VtsAidlKeyMintTargetTest
Bug: 192222727
Change-Id: I9f785579e38dbbc45e1e4f993a991ff38b59b6cd
Merged-In: I9f785579e38dbbc45e1e4f993a991ff38b59b6cd