Only insert CREATION_DATETIME for KeyMaster
If the reference implementation is acting as a Key*Mint* instance,
then CREATION_DATETIME should only be echoed back if it was provided
at key generation/import; it should not be created from scratch.

Update the software-only version of the code to do this.

(This code is not currently used by Cuttlefish, which is why this was
not spotted.)

Bug: 204795916
Test: VtsAidlKeyMintTargetTest
Change-Id: I133faa3f8d7a6e035449d0d56ab1e75dd63480d9