Pass verified boot info to software keymint
Add SetVerifiedBootInfo to KeymasterContext, allowing information about
the AVB status to be sent to keymint.

Add verious setters to RemoteProvisioningContext that mirror those in
KeymasterContext, as RemoteProvisioningContext now needs the same
data for building DeviceInfo.

Add ConfigureVerifiedBootInfoRequest for passing the verified boot
info over the wire to a keymint implementation. This includes new
serialization to make it easier to put C++ collections in messages.
Logging errors on deserialization required a small change to logger.h,
as the macro names conflict with identical ones in trusty.

Update pure software keymint with verified boot info.

Test: keymaster_tests
Test: VtsHalRemotelyProvisionedComponentTargetTest
Change-Id: Icd56c6616e5dcd2d2f31862e1444c306f393609b