Add a missing LazyInitProdBcc call.
With the introduction of the generateCertificateRequestV2 call, there
was a new code path introduce in android_keymaster to handle this
additional HAL method.

The underlying code this HAL called in the pure software provisioning
context did not also trigger generation of the BCC and device private
key for the new HAL method.

This is now fixed by initializing the Bcc in BuildCsr

Fixes: 270522174
Test: N/A
Change-Id: Ic49008d832e6b1fdc846817a8e7dcda07657983a